fix signed integer comparisons

+++ src/RadMarshakAsymptotic/test_radiation_marshak_asymptotic.cpp
@@ -304,7 +304,7 @@ auto problem_main() -> int
 
 	double err_norm = 0.;
 	double sol_norm = 0.;
-	for (int i = 0; i < xs_exact.size(); ++i) {
+	for (size_t i = 0; i < xs_exact.size(); ++i) {
 		err_norm += std::abs(Tmat_interp[i] - Tmat_exact[i]);
 		sol_norm += std::abs(Tmat_exact[i]);
 	}
