+++ src/RadPulse/test_radiation_pulse.cpp
@@ -196,7 +196,7 @@ auto problem_main() -> int
 	// compute error norm
 	double err_norm = 0.;
 	double sol_norm = 0.;
-	for (int i = 0; i < xs.size(); ++i) {
+	for (size_t i = 0; i < xs.size(); ++i) {
 		err_norm += std::abs(Trad[i] - Trad_exact[i]);
 		sol_norm += std::abs(Trad_exact[i]);
 	}
